Add files via upload
Object tracking in computer vision is following and keeping a record of the position of any object upon its change in movement in the video. Multiple objects can also be tracked simultaneously. The main goal of object tracking is to maintain the identity and positioning of the object as it moves through the video, despite variations in appearance, size, orientation, and lighting, and releasing the object tracking once the object goes out of the video frame.